upgrade Kotlin to 1.2 add Kotlin DSL for querying and pagination add Kotlin extension function to allow KProperty references in lieu of string property references fix pagination output: one extra row was included test/fix File caching

+++ src/main/kotlin/sqlbuilder/FileStrategy.kt
@@ -1,44 +1,54 @@
 package sqlbuilder
 
-import org.slf4j.LoggerFactory
+import sqlbuilder.exceptions.CacheException
 import java.io.File
 import java.io.FileInputStream
 import java.io.FileOutputStream
 import java.io.ObjectInputStream
 import java.io.ObjectOutputStream
+import java.io.Serializable
 
 /**
  * Caching strategy that use Serialisation to store any query to a single file.
- * This means the user must specify a different file for every dataset.
+ * This means the user must specify a different file for every query and parameter combination.
  *
  * @author Laurent Van der Linden
  */
-class FileStrategy(private var file: File) : CacheStrategy {
-    private val logger = LoggerFactory.getLogger(javaClass)!!
-
+class FileStrategy(private val file: File) : CacheStrategy {
     override fun get(query: CacheableQuery): Any? {
-
-        if (file.isFile) {
+        if (file.isFile && file.length() > 0) {
             try {
                 return FileInputStream(file).use { fis ->
                     ObjectInputStream(fis).readObject()
                 }
             } catch (e: Exception) {
-                logger.warn("failed to read cached value from file ${file.absolutePath}", e)
-                file.delete()
+                e.printStackTrace()
+                try {
+                    file.delete()
+                } finally {
+                    throw CacheException("failed to read cached value from file ${file.absolutePath}", e)
+                }
             }
-
         }
         return null
     }
 
     override fun put(query: CacheableQuery, result: Any) {
-        try {
-            ObjectOutputStream(FileOutputStream(file)).use { oos ->
-                oos.writeObject(result)
+        if (result is Serializable) {
+            try {
+                ObjectOutputStream(FileOutputStream(file)).use { oos ->
+                    oos.writeObject(result)
+                }
+            } catch (e: Exception) {
+                try {
+                    file.delete()
+                } finally {
+                    throw CacheException("failed to cache value to file ${file.absolutePath}", e)
+                }
+            } finally {
             }
-        } catch (e: Exception) {
-            logger.warn("failed to cache value to file", e)
+        } else {
+            throw CacheException("when using a FileStrategy cache, all query results must be java.io.Serializable")
         }
     }
 }
